requirements.txt
Flask==3.0.0
flask-cors==4.0.0
cloudinary==1.38.0
gunicorn==21.2.0
Werkzeug==3.0.1
python-dotenv==1.0.1 